+++ test/test_feature_base.py
@@ -9,7 +9,7 @@ from ase.ga.data import DataConnection
 
 from atoml import __path__ as atoml_path
 from atoml.fingerprint.base import BaseGenerator
-from atoml.utilities.neighborlist import atoms_neighborlist
+from atoml.utilities.neighborlist import ase_neighborlist
 
 atoml_path = '/'.join(atoml_path[0].split('/')[:-1])
 
@@ -23,7 +23,7 @@ class TestBaseGenerator(unittest.TestCase):
         all_cand = gadb.get_all_relaxed_candidates()
 
         f = BaseGenerator()
-        nl = atoms_neighborlist(all_cand[0])
+        nl = ase_neighborlist(all_cand[0])
         assert f.get_neighborlist(all_cand[0]) == nl
 
         pos = all_cand[0].get_positions()