+++ CHANGELOG.md
@@ -1,7 +1,16 @@
-# Version 0.3.2 (March 2018)
+# dev
 
--   Voronoi and prototype feature generation added.
--   Magpie api added.
+-   Magpie api added, allows for Voronoi and prototype feature generation.
+-   A genetic algorithm added for feature optimization.
+-   Parallelism updated to be compatable with Python2.
+-   Added in better neighborlist generation.
+    -   Updated wrapper for ase neighborlist.
+    -   Updated AtoML neighborlist generator.
+-   Added basic NetworkX api.
+-   Added some general functions to clean data and build a GP.
+-   Added a test for dependencies. Will raise a warning in the CI if things get out of date.
+-   Added a custom docker image for the tests. This is compiled in the `setup/` directory in root.
+-   Clean up some bits of code, fix some bugs.
 
 # Version 0.3.1 (February 2018)